Separate the remaining default handlers into separate modules
This change moves the ignore and copy handlers into separate files and introduces a test for the ignore handler.
